requirements.txt
playwright==1.52.0
beautifulsoup4==4.12.3
google-generativeai==0.8.5
supabase==2.15.1
python-dotenv==1.0.1
APScheduler==3.10.4
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
